[ADD] Logic for the default approval users set.
