+++ document_page/document_page.py
@@ -98,23 +98,11 @@ class document_page(models.Model):
         readonly=True
     )
 
-    revision = fields.Integer(
-        string='Revision number',
-        compute='_get_revision',
-        readonly=True)
-
     document_type = fields.Selection(
         string='Document type',
         selection=[('internal', 'Internal'),
                    ('external', 'External')],
         default='internal')
-
-    def _get_revision(self):
-        for page in self:
-            revision = []
-            [revision.append(history.id) for history in page.history_ids
-             if history.state == 'approved']
-            page.revision = len(revision)
 
     def _get_page_index(self, page, link=True):
         index = []
