+++ document_page/document_page_view.xml
@@ -31,6 +31,7 @@
             <field name="arch" type="xml">
                 <tree string="Document Page">
                     <field name="name"/>
+                    <field name="document_type"/>
                     <field name="parent_id"/>
                     <field name="create_uid" invisible="1"/>
                     <field name="write_uid"/>
@@ -52,10 +53,7 @@
                             <field name="parent_id"
                                    string="Category"
                                    context="{'default_type':'category'}"/>
-                            <field name="revision" />
-                        </group>
-                        <group>
-                             <field name="document_type"
+                            <field name="document_type"
                                 context="{'default_type':'internal'}"/>
                         </group>
                         <group>
